# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MyCarQR API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /alerts:
    get:
      operationId: getAlerts
      tags: [alerts]
      summary: List user's scan alerts
      security:
        - bearerAuth: []
      parameters:
        - name: vehicleId
          in: query
          schema:
            type: integer
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScanAlert"
  /alerts/{alertId}/read:
    post:
      operationId: markAlertRead
      tags: [alerts]
      summary: Mark alert as read
      security:
        - bearerAuth: []
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanAlert"
  /alerts/mark-all-read:
    post:
      operationId: markAllAlertsRead
      tags: [alerts]
      summary: Mark all alerts as read
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    type: integer
                required: [updated]
  /documents:
    get:
      operationId: getDocuments
      tags: [documents]
      summary: List vehicle documents/reminders
      security:
        - bearerAuth: []
      parameters:
        - name: vehicleId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VehicleDocument"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Add a document reminder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentBody"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleDocument"
  /documents/{documentId}:
    put:
      operationId: updateDocument
      tags: [documents]
      summary: Update a document reminder
      security:
        - bearerAuth: []
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentBody"
      responses:
        "200":
          description: Updated document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleDocument"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document reminder
      security:
        - bearerAuth: []
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /lost-items:
    get:
      operationId: getLostItems
      tags: [alerts]
      summary: List lost item reports for user's vehicles
      security:
        - bearerAuth: []
      parameters:
        - name: vehicleId
          in: query
          schema:
            type: integer
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of lost item reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LostItem"
  /lost-items/{itemId}/read:
    post:
      operationId: markLostItemRead
      tags: [alerts]
      summary: Mark lost item report as read
      security:
        - bearerAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LostItem"
...
  /payment-requests:
    get:
      operationId: getMyPaymentRequests
      tags: [payments]
      summary: Get own payment requests
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of payment requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentRequest"
    post:
      operationId: submitPaymentRequest
      tags: [payments]
      summary: Submit a new payment request with screenshot
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitPaymentBody"
      responses:
        "201":
          description: Payment request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentRequest"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ScanAlert:
      type: object
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        vehicleNumber:
          type: string
        alertType:
          type: string
        message:
          type: string
        scannerLocation:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - vehicleId
        - alertType
        - isRead
        - createdAt
...
    VehicleDocument:
      type: object
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        vehicleNumber:
          type: string
        documentType:
          type: string
          enum: [insurance, pollution, registration, service]
        expiryDate:
          type: string
        status:
          type: string
          enum: [up_to_date, expiring_soon, expired]
        notes:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - vehicleId
        - documentType
        - expiryDate
        - status
        - createdAt
        - updatedAt
    CreateDocumentBody:
      type: object
      properties:
        vehicleId:
          type: integer
        documentType:
          type: string
          enum: [insurance, pollution, registration, service]
        expiryDate:
          type: string
        notes:
          type: string
      required:
        - vehicleId
        - documentType
        - expiryDate
...
    LostItem:
      type: object
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        vehicleNumber:
          type: string
        message:
          type: string
        photos:
          type: array
          items:
            type: string
        latitude:
          type: string
        longitude:
          type: string
        locationLabel:
          type: string
        finderContact:
          type: string
        reportedAt:
          type: string
        isRead:
          type: boolean
      required:
        - id
        - vehicleId
        - message
        - photos
        - reportedAt
        - isRead
...
    PaymentRequest:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        email:
          type: string
        status:
          type: string
        planType:
          type: string
        amount:
          type: integer
        screenshotBase64:
          type: string
        adminNote:
          type: string
        createdAt:
          type: string
        reviewedAt:
          type: string
        expiresAt:
          type: string
      required:
        - id
        - userId
        - status
        - planType
        - amount
        - createdAt
    SubmitPaymentBody:
      type: object
      properties:
        planType:
          type: string
        amount:
          type: integer
        screenshotBase64:
          type: string
      required:
        - planType
        - amount
...